+++ pjax.js
@@ -1,4 +1,39 @@
 "use strict";
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
+var __generator = (this && this.__generator) || function (thisArg, body) {
+    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
+    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
+    function verb(n) { return function (v) { return step([n, v]); }; }
+    function step(op) {
+        if (f) throw new TypeError("Generator is already executing.");
+        while (_) try {
+            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
+            if (y = 0, t) op = [op[0] & 2, t.value];
+            switch (op[0]) {
+                case 0: case 1: t = op; break;
+                case 4: _.label++; return { value: op[1], done: false };
+                case 5: _.label++; y = op[1]; op = [0]; continue;
+                case 7: op = _.ops.pop(); _.trys.pop(); continue;
+                default:
+                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
+                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
+                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
+                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
+                    if (t[2]) _.ops.pop();
+                    _.trys.pop(); continue;
+            }
+            op = body.call(thisArg, _);
+        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
+        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
+    }
+};
 Object.defineProperty(exports, "__esModule", { value: true });
 var parse_options_1 = require("./lib/parse-options");
 var trigger_1 = require("./lib/events/trigger");
@@ -109,7 +144,7 @@ var Pjax = (function () {
         }
         this.finalize();
     };
-    Pjax.prototype.switchSelectors = function (selectors, tempDocument, currentDocument) {
+    Pjax.prototype.switchSelectors = function (selectors, tempDocument) {
         if (tempDocument === null) {
             if (this.options.debug) {
                 console.log('%c[Pjax] ' + ("%ctemporary document was null, telling the browser to load " + ((this._cache !== null) ? this._cache.url : this._response.url)), 'color:#f3ff35', 'color:#eee');
@@ -122,10 +157,9 @@ var Pjax = (function () {
             }
         }
         var switchQueue = [];
-        var contiansScripts = false;
         for (var i = 0; i < selectors.length; i++) {
             var newContainers = Array.from(tempDocument.querySelectorAll(selectors[i]));
-            var currentContainers = Array.from(currentDocument.querySelectorAll(selectors[i]));
+            var currentContainers = Array.from(document.querySelectorAll(selectors[i]));
             if (this.options.debug) {
                 console.log('%c[Pjax] ' + ("%cswapping content from " + selectors[i]), 'color:#f3ff35', 'color:#eee');
             }
@@ -137,10 +171,6 @@ var Pjax = (function () {
                 return;
             }
             for (var k = 0; k < newContainers.length; k++) {
-                var scripts = Array.from(newContainers[k].querySelectorAll('script'));
-                if (scripts.length > 0) {
-                    contiansScripts = true;
-                }
                 var newContainer = newContainers[k];
                 var currentContainer = currentContainers[k];
                 var switchObject = {
@@ -153,13 +183,6 @@ var Pjax = (function () {
         if (switchQueue.length === 0) {
             if (this.options.debug) {
                 console.log('%c[Pjax] ' + "%ccouldn't find anything to switch", 'color:#f3ff35', 'color:#eee');
-            }
-            this.lastChance(this._response.url);
-            return;
-        }
-        if (contiansScripts) {
-            if (this.options.debug) {
-                console.log('%c[Pjax] ' + "%cthe new page contains scripts", 'color:#f3ff35', 'color:#eee');
             }
             this.lastChance(this._response.url);
             return;
@@ -198,7 +221,10 @@ var Pjax = (function () {
         }
         clear_active_1.default();
         state_manager_1.default.doReplace(window.location.href, document.title);
-        this.switchSelectors(this.options.selectors, this._cache.document, document);
+        if (this.options.importScripts) {
+            this.handleScripts(this._cache.document);
+        }
+        this.switchSelectors(this.options.selectors, this._cache.document);
     };
     Pjax.prototype.parseContent = function (responseText) {
         var tempDocument = document.implementation.createHTMLDocument('pjax-temp-document');
@@ -235,7 +261,10 @@ var Pjax = (function () {
         if (tempDocument instanceof HTMLDocument) {
             clear_active_1.default();
             state_manager_1.default.doReplace(window.location.href, document.title);
-            this.switchSelectors(this.options.selectors, tempDocument, document);
+            if (this.options.importScripts) {
+                this.handleScripts(tempDocument);
+            }
+            this.switchSelectors(this.options.selectors, tempDocument);
         }
         else {
             if (this.options.debug) {
@@ -244,6 +273,56 @@ var Pjax = (function () {
             trigger_1.default(document, ['pjax:error']);
             this.lastChance(this._response.url);
             return;
+        }
+    };
+    Pjax.prototype.handleScripts = function (newDocument) {
+        var _this = this;
+        var newScripts = Array.from(newDocument.querySelectorAll('script'));
+        var currentScripts = Array.from(document.querySelectorAll('script'));
+        var scriptsToAppend = [];
+        newScripts.forEach(function (newScript) {
+            var appendScript = true;
+            currentScripts.forEach(function (currentScript) {
+                if (newScript.src === currentScript.src
+                    || newScript.src === currentScript.dataset.source
+                    || "" + window.location.origin + window.location.pathname + newScript.src === currentScript.dataset.source
+                    || "" + window.location.origin + window.location.pathname + newScript.src === currentScript.src) {
+                    appendScript = false;
+                }
+            });
+            if (appendScript) {
+                scriptsToAppend.push(newScript);
+            }
+        });
+        if (scriptsToAppend.length) {
+            scriptsToAppend.forEach(function (script) {
+                if (script.src === '') {
+                    var newScript = document.createElement('script');
+                    newScript.dataset.source = _this._response.url;
+                    newScript.innerHTML = script.innerHTML;
+                    document.body.appendChild(newScript);
+                }
+                else {
+                    (function () { return __awaiter(_this, void 0, void 0, function () {
+                        var response, responseText, newScript;
+                        return __generator(this, function (_a) {
+                            switch (_a.label) {
+                                case 0: return [4, fetch(script.src)];
+                                case 1:
+                                    response = _a.sent();
+                                    return [4, response.text()];
+                                case 2:
+                                    responseText = _a.sent();
+                                    newScript = document.createElement('script');
+                                    newScript.dataset.source = script.src;
+                                    newScript.innerHTML = responseText;
+                                    document.body.appendChild(newScript);
+                                    return [2];
+                            }
+                        });
+                    }); })();
+                }
+            });
         }
     };
     Pjax.prototype.handleResponse = function (response) {