+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Adds: when `importScripts` is set to `true` (default) any script elements will be fetched and appended to the `document.body`
 - Adds: when `importCSS` is set to `true` (default) any CSS (in `HEAD` or via `<link>`) will be fetched and appended
 
+### Removes
+
+- Removes: no longer calls `lastChance()` when the new document contains `<script>` elements
+
 # 1.3.0 - 2019-03-27
 
 ### Adds